GraphQL 테스트 책임을 기능별로 정리한다
