+++ apps/api/package.json
@@ -7,7 +7,7 @@
     "lint:schema": "NODE_ENV=production node --import tsx scripts/check-schema.ts",
     "lint:tsc": "tsc --noEmit",
     "start": "node --import tsx src/index.ts",
-    "test": "pnpm lint:tsc && pnpm test:unit && pnpm test:integration",
+    "test": "pnpm lint:schema && pnpm lint:tsc && pnpm test:unit && pnpm test:integration",
     "test:integration": "cd ../.. && node scripts/test-db.mjs run -- pnpm --filter @kosmo/api test:integration:database",
     "test:integration:database": "cd ../.. && pnpm db:test:push && pnpm --filter @kosmo/api exec node --import tsx --test --test-concurrency=1 'tests/integration/**/*.test.ts'",
     "test:unit": "NODE_ENV=production node --import tsx --test 'src/**/*.test.ts'"
