+++ apps/api/src/graphql/resolvers/post/mutation/create.test.ts
@@ -1,0 +1,27 @@
+import assert from 'node:assert/strict';
+import test from 'node:test';
+import { graphql } from 'graphql';
+import { schema } from '@/graphql/schema';
+
+test('rejects empty and over-500-character Plain Text before creating a post', async () => {
+  for (const [bodyText, message] of [
+    [' \n ', '본문을 입력해주세요.'],
+    ['가'.repeat(501), '본문은 500자까지 작성할 수 있어요.'],
+  ] as const) {
+    const result = await graphql({
+      schema,
+      source: `
+        mutation CreatePost($input: CreatePostInput!) {
+          createPost(input: $input) {
+            post { id }
+          }
+        }
+      `,
+      variableValues: { input: { bodyText, visibility: 'UNLISTED' } },
+      contextValue: { session: { profileId: '00000000-0000-8000-8000-000000000001' } },
+    });
+
+    assert.equal(result.data == null, true);
+    assert.equal(result.errors?.[0]?.message, message);
+  }
+});
