--- apps/api/src/graphql/schema.test.ts
@@ -1,561 +1,0 @@
-import assert from 'node:assert/strict';
-import test from 'node:test';
-import { encodeGlobalId } from '@kosmo/core/global-id';
-import { graphql, isInputObjectType, isInterfaceType, isObjectType, isUnionType } from 'graphql';
-import { notificationNodeType } from './resolvers/notification/ref';
-import { schema } from './schema';
-
-test('exposes upload URL issuance with a minimal Media state contract', () => {
-  const mutation = schema.getMutationType();
-  const payload = schema.getType('IssueMediaUploadUrlPayload');
-  const media = schema.getType('Media');
-
-  const field = mutation?.getFields().issueMediaUploadUrl;
-  assert.equal(String(field?.type), 'IssueMediaUploadUrlPayload!');
-  assert.deepEqual(field?.args, []);
-
-  assert.ok(isObjectType(payload));
-  assert.deepEqual(Object.keys(payload.getFields()).sort(), ['expiresAt', 'media', 'uploadUrl']);
-  assert.equal(String(payload.getFields().media.type), 'Media!');
-  assert.equal(String(payload.getFields().uploadUrl.type), 'String!');
-  assert.equal(String(payload.getFields().expiresAt.type), 'DateTime!');
-
-  assert.ok(isObjectType(media));
-  assert.deepEqual(Object.keys(media.getFields()).sort(), ['id', 'readyAt', 'state']);
-  assert.equal(String(media.getFields().readyAt.type), 'DateTime');
-  assert.equal(String(media.getFields().state.type), 'MediaState!');
-  assert.equal(media.getFields().storageReference, undefined);
-});
-
-test('exposes idempotent Local Media upload completion without a storage reference', () => {
-  const mutation = schema.getMutationType();
-  const field = mutation?.getFields().completeMediaUpload;
-  const input = schema.getType('CompleteMediaUploadInput');
-  const payload = schema.getType('CompleteMediaUploadPayload');
-
-  assert.equal(String(field?.type), 'CompleteMediaUploadPayload!');
-  assert.equal(field?.args[0]?.name, 'input');
-  assert.equal(String(field?.args[0]?.type), 'CompleteMediaUploadInput!');
-
-  assert.ok(isInputObjectType(input));
-  assert.deepEqual(Object.keys(input.getFields()), ['id']);
-  assert.equal(String(input.getFields().id.type), 'ID!');
-
-  assert.ok(isObjectType(payload));
-  assert.deepEqual(Object.keys(payload.getFields()), ['media']);
-  assert.equal(String(payload.getFields().media.type), 'Media!');
-  assert.equal(payload.getFields().storageReference, undefined);
-});
-
-test('requires an authenticated selected Profile for Media upload mutations', async () => {
-  const mediaId = encodeGlobalId('Media', '00000000-0000-8000-8000-000000000001');
-  for (const source of [
-    `mutation { issueMediaUploadUrl { uploadUrl } }`,
-    `mutation { completeMediaUpload(input: { id: "${mediaId}" }) { media { id } } }`,
-  ]) {
-    for (const contextValue of [
-      {},
-      { session: { id: 'session', accountId: 'account', profileId: null } },
-    ]) {
-      const result = await graphql({ schema, source, contextValue });
-
-      assert.equal(result.data, null);
-      assert.match(result.errors?.[0]?.message ?? '', /Not authorized/);
-    }
-  }
-});
-
-test('Post는 nullable Repost Source를 제공한다', () => {
-  const post = schema.getType('Post');
-
-  assert.ok(isObjectType(post));
-  assert.equal(String(post.getFields().repostSource?.type), 'Post');
-});
-
-test('exposes partial Profile handle search without changing exact lookup', () => {
-  const query = schema.getQueryType();
-
-  assert.ok(query);
-  assert.equal(String(query.getFields().profileByHandle?.type), 'Profile');
-  assert.equal(String(query.getFields().searchProfiles?.type), 'ProfileConnection!');
-  assert.deepEqual(
-    query.getFields().searchProfiles?.args.map(({ name }) => name),
-    ['after', 'before', 'first', 'last', 'query'],
-  );
-});
-
-test('exposes viewer-independent Repost count and selected Profile Repost on Post', () => {
-  const post = schema.getType('Post');
-
-  assert.ok(isObjectType(post));
-  assert.equal(String(post.getFields().repostCount?.type), 'Int!');
-  assert.equal(String(post.getFields().viewerRepost?.type), 'Post');
-});
-
-test('exposes viewer-independent Reaction counts on Post', () => {
-  const post = schema.getType('Post');
-  const reactionCount = schema.getType('ReactionCount');
-
-  assert.ok(isObjectType(post));
-  assert.equal(String(post.getFields().reactionCounts?.type), '[ReactionCount!]!');
-  assert.deepEqual(post.getFields().reactionCounts?.args, []);
-
-  assert.ok(isObjectType(reactionCount));
-  assert.equal(String(reactionCount.getFields().type?.type), 'String!');
-  assert.equal(String(reactionCount.getFields().count?.type), 'Int!');
-  assert.deepEqual(Object.keys(reactionCount.getFields()).sort(), ['count', 'type']);
-});
-
-test('exposes the versioned PostContent document and Plain Text composer contract', () => {
-  const postContent = schema.getType('PostContent');
-  const createPostInput = schema.getType('CreatePostInput');
-
-  assert.ok(isObjectType(postContent));
-  assert.ok(postContent.getFields().document);
-  assert.ok(postContent.getFields().bodyText);
-  assert.ok(postContent.getFields().contentWarning);
-  assert.equal(postContent.getFields().spoilerText, undefined);
-  assert.equal(postContent.getFields().bodyJson, undefined);
-
-  assert.ok(isInputObjectType(createPostInput));
-  assert.equal(createPostInput.getFields().content, undefined);
-  assert.equal(String(createPostInput.getFields().bodyText?.type), 'String!');
-  assert.equal(String(createPostInput.getFields().replyParentId?.type), 'ID');
-  assert.equal(schema.getType('TipTapDocument'), undefined);
-  assert.equal(schema.getType('PostContentBody'), undefined);
-  assert.equal(String(schema.getType('PostContentDocument')), 'PostContentDocument');
-});
-
-test('exposes Reply Parent through the existing nullable Post Node contract', () => {
-  const post = schema.getType('Post');
-
-  assert.ok(isObjectType(post));
-  assert.equal(String(post.getFields().replyParent?.type), 'Post');
-  assert.equal(post.getFields().replyParentId, undefined);
-});
-
-test('exposes Reply ancestors as a non-null Post list without pagination', () => {
-  const post = schema.getType('Post');
-
-  assert.ok(isObjectType(post));
-  assert.equal(String(post.getFields().replyAncestors?.type), '[Post!]!');
-  assert.deepEqual(post.getFields().replyAncestors?.args, []);
-});
-
-test('exposes Reply descendants through the shared Post connection', () => {
-  const post = schema.getType('Post');
-
-  assert.ok(isObjectType(post));
-  const field = post.getFields().replyDescendants;
-  assert.equal(String(field?.type), 'PostConnection!');
-  assert.deepEqual(
-    field?.args.map(({ name }) => name),
-    ['after', 'before', 'first', 'last'],
-  );
-});
-
-test('follow mutation payloads expose both updated profiles', () => {
-  const followPayload = schema.getType('FollowProfilePayload');
-  const unfollowPayload = schema.getType('UnfollowProfilePayload');
-
-  assert.ok(isObjectType(followPayload));
-  assert.ok(followPayload.getFields().followerProfile);
-  assert.ok(followPayload.getFields().followeeProfile);
-  assert.equal(followPayload.getFields().profile, undefined);
-
-  assert.ok(isObjectType(unfollowPayload));
-  assert.ok(unfollowPayload.getFields().followerProfile);
-  assert.ok(unfollowPayload.getFields().followeeProfile);
-  assert.equal(String(unfollowPayload.getFields().followerProfile?.type), 'Profile!');
-  assert.equal(String(unfollowPayload.getFields().followeeProfile?.type), 'Profile');
-  assert.equal(unfollowPayload.getFields().profile, undefined);
-});
-
-test('exposes the minimal idempotent Reaction add contract', () => {
-  const mutation = schema.getMutationType();
-  const input = schema.getType('AddReactionInput');
-  const payload = schema.getType('AddReactionPayload');
-  const reaction = schema.getType('Reaction');
-
-  assert.equal(String(mutation?.getFields().addReaction?.type), 'AddReactionPayload!');
-  assert.ok(isInputObjectType(input));
-  assert.equal(String(input.getFields().postId.type), 'ID!');
-  assert.equal(String(input.getFields().type.type), 'String!');
-  assert.ok(isObjectType(payload));
-  assert.equal(String(payload.getFields().reaction.type), 'Reaction!');
-  assert.equal(payload.getFields().created, undefined);
-  assert.ok(isObjectType(reaction));
-  assert.deepEqual(
-    reaction.getInterfaces().map(({ name }) => name),
-    ['Node'],
-  );
-  assert.equal(String(reaction.getFields().type.type), 'String!');
-  assert.equal(String(reaction.getFields().createdAt.type), 'DateTime!');
-});
-
-test('exposes the idempotent Repost creation contract', () => {
-  const mutation = schema.getMutationType();
-  const input = schema.getType('RepostPostInput');
-  const payload = schema.getType('RepostPostPayload');
-
-  assert.equal(String(mutation?.getFields().repostPost?.type), 'RepostPostPayload!');
-  assert.ok(isInputObjectType(input));
-  assert.equal(String(input.getFields().sourceId.type), 'ID!');
-  assert.ok(isObjectType(payload));
-  assert.equal(String(payload.getFields().repost.type), 'Post!');
-  assert.equal(payload.getFields().created, undefined);
-});
-
-test('exposes the ID-based idempotent Post delete contract', () => {
-  const mutation = schema.getMutationType();
-  const input = schema.getType('DeletePostInput');
-  const payload = schema.getType('DeletePostPayload');
-
-  assert.equal(String(mutation?.getFields().deletePost?.type), 'DeletePostPayload!');
-  assert.ok(isInputObjectType(input));
-  assert.equal(String(input.getFields().id.type), 'ID!');
-  assert.ok(isObjectType(payload));
-  assert.equal(String(payload.getFields().postId.type), 'ID!');
-  assert.equal(payload.getFields().post, undefined);
-  assert.equal(payload.getFields().deleted, undefined);
-});
-
-test('exposes the Bookmark mutation and relationship contract', () => {
-  const mutation = schema.getMutationType();
-  const createInput = schema.getType('CreateBookmarkInput');
-  const createPayload = schema.getType('CreateBookmarkPayload');
-  const deleteInput = schema.getType('DeleteBookmarkInput');
-  const deletePayload = schema.getType('DeleteBookmarkPayload');
-  const bookmark = schema.getType('Bookmark');
-  const post = schema.getType('Post');
-
-  assert.equal(String(mutation?.getFields().createBookmark?.type), 'CreateBookmarkPayload!');
-  assert.ok(isInputObjectType(createInput));
-  assert.equal(String(createInput.getFields().postId.type), 'ID!');
-  assert.ok(isObjectType(createPayload));
-  assert.equal(String(createPayload.getFields().bookmark.type), 'Bookmark!');
-  assert.equal(String(mutation?.getFields().deleteBookmark?.type), 'DeleteBookmarkPayload!');
-  assert.ok(isInputObjectType(deleteInput));
-  assert.equal(String(deleteInput.getFields().id.type), 'ID!');
-  assert.ok(isObjectType(deletePayload));
-  assert.equal(String(deletePayload.getFields().bookmarkId.type), 'ID');
-  assert.equal(String(deletePayload.getFields().post.type), 'Post');
-  assert.ok(isObjectType(bookmark));
-  assert.equal(String(bookmark.getFields().profile.type), 'Profile!');
-  assert.equal(String(bookmark.getFields().post.type), 'Post');
-  assert.equal(String(bookmark.getFields().createdAt.type), 'DateTime!');
-  assert.ok(isObjectType(post));
-  assert.equal(String(post.getFields().viewerBookmark.type), 'Bookmark');
-});
-
-test('exposes selected Profile Reactions and the Post/Type delete contract', () => {
-  const mutation = schema.getMutationType();
-  const input = schema.getType('DeleteReactionInput');
-  const payload = schema.getType('DeleteReactionPayload');
-  const post = schema.getType('Post');
-
-  assert.equal(String(mutation?.getFields().deleteReaction?.type), 'DeleteReactionPayload!');
-  assert.ok(isInputObjectType(input));
-  assert.equal(String(input.getFields().postId.type), 'ID!');
-  assert.equal(String(input.getFields().type.type), 'String!');
-  assert.ok(isObjectType(payload));
-  assert.equal(String(payload.getFields().reactionId.type), 'ID');
-  assert.equal(String(payload.getFields().post.type), 'Post');
-  assert.equal(payload.getFields().reaction, undefined);
-  assert.equal(payload.getFields().deleted, undefined);
-  assert.ok(isObjectType(post));
-  assert.equal(String(post.getFields().viewerReactions.type), '[Reaction!]!');
-});
-
-test('exposes Reaction Profiles as the shared Profile connection without Reaction metadata', () => {
-  const post = schema.getType('Post');
-  const connection = schema.getType('ProfileConnection');
-  const edge = schema.getType('ProfileConnectionEdge');
-
-  assert.ok(isObjectType(post));
-  const field = post.getFields().reactionProfiles;
-  assert.equal(String(field?.type), 'ProfileConnection!');
-  assert.equal(String(field?.args.find(({ name }) => name === 'type')?.type), 'String!');
-
-  assert.ok(isObjectType(connection));
-  assert.ok(isObjectType(edge));
-  assert.equal(String(edge.getFields().node.type), 'Profile!');
-  assert.equal(edge.getFields().reaction, undefined);
-  assert.equal(edge.getFields().reactionId, undefined);
-  assert.equal(edge.getFields().reactedAt, undefined);
-});
-
-test('exposes the profile follow request lifecycle contract', () => {
-  const profile = schema.getType('Profile');
-  const viewerState = schema.getType('ProfileViewerState');
-  const request = schema.getType('ProfileFollowRequest');
-  const result = schema.getType('ProfileFollowResult');
-  const followPayload = schema.getType('FollowProfilePayload');
-  const approvePayload = schema.getType('ApproveProfileFollowRequestPayload');
-  const rejectPayload = schema.getType('RejectProfileFollowRequestPayload');
-  const cancelPayload = schema.getType('CancelProfileFollowRequestPayload');
-
-  assert.ok(isObjectType(profile));
-  assert.equal(profile.getFields().viewerFollowRequest, undefined);
-  assert.equal(profile.getFields().viewerFollow, undefined);
-  assert.equal(
-    String(profile.getFields().incomingProfileFollowRequests?.type),
-    'ProfileIncomingProfileFollowRequestsConnection',
-  );
-  assert.equal(
-    String(profile.getFields().outgoingProfileFollowRequests?.type),
-    'ProfileOutgoingProfileFollowRequestsConnection',
-  );
-
-  assert.ok(isObjectType(request));
-  assert.equal(String(request.getFields().follower?.type), 'Profile');
-  assert.equal(String(request.getFields().followee?.type), 'Profile');
-
-  assert.ok(isObjectType(viewerState));
-  assert.equal(String(viewerState.getFields().follow?.type), 'ProfileFollow');
-  assert.equal(String(viewerState.getFields().followRequest?.type), 'ProfileFollowRequest');
-
-  assert.ok(isUnionType(result));
-  assert.deepEqual(
-    result
-      .getTypes()
-      .map(({ name }) => name)
-      .sort(),
-    ['ProfileFollow', 'ProfileFollowRequest'],
-  );
-
-  assert.ok(isObjectType(followPayload));
-  assert.equal(String(followPayload.getFields().result?.type), 'ProfileFollowResult!');
-  assert.equal(followPayload.getFields().profileFollow, undefined);
-
-  assert.ok(isObjectType(approvePayload));
-  assert.equal(String(approvePayload.getFields().profileFollow?.type), 'ProfileFollow!');
-  assert.equal(String(approvePayload.getFields().profileFollowRequestId?.type), 'ID!');
-  assert.equal(String(approvePayload.getFields().followerProfile?.type), 'Profile!');
-  assert.equal(String(approvePayload.getFields().followeeProfile?.type), 'Profile!');
-
-  assert.ok(isObjectType(rejectPayload));
-  assert.equal(String(rejectPayload.getFields().profileFollowRequestId?.type), 'ID!');
-  assert.equal(String(rejectPayload.getFields().followeeProfile?.type), 'Profile!');
-  assert.equal(rejectPayload.getFields().followerProfile, undefined);
-  assert.equal(rejectPayload.getFields().profileFollowRequest, undefined);
-
-  assert.ok(isObjectType(cancelPayload));
-  assert.equal(String(cancelPayload.getFields().profileFollowRequestId?.type), 'ID!');
-  assert.equal(String(cancelPayload.getFields().followerProfile?.type), 'Profile!');
-  assert.equal(cancelPayload.getFields().followeeProfile, undefined);
-  assert.equal(cancelPayload.getFields().profileFollowRequest, undefined);
-});
-
-test('rejects empty and over-500-character Plain Text before creating a post', async () => {
-  for (const [bodyText, message] of [
-    [' \n ', '본문을 입력해주세요.'],
-    ['가'.repeat(501), '본문은 500자까지 작성할 수 있어요.'],
-  ] as const) {
-    const result = await graphql({
-      schema,
-      source: `
-        mutation CreatePost($input: CreatePostInput!) {
-          createPost(input: $input) {
-            post { id }
-          }
-        }
-      `,
-      variableValues: { input: { bodyText, visibility: 'UNLISTED' } },
-      contextValue: { session: { profileId: '00000000-0000-8000-8000-000000000001' } },
-    });
-
-    assert.equal(result.data == null, true);
-    assert.equal(result.errors?.[0]?.message, message);
-  }
-});
-
-test('exposes Notification interface and concrete source types without raw storage fields', () => {
-  const notification = schema.getType('Notification');
-  const followNotification = schema.getType('FollowNotification');
-  const reactionNotification = schema.getType('ReactionNotification');
-  const replyNotification = schema.getType('ReplyNotification');
-  const repostNotification = schema.getType('RepostNotification');
-  const profile = schema.getType('Profile');
-
-  assert.ok(isObjectType(followNotification));
-  assert.ok(isObjectType(profile));
-  assert.deepEqual(
-    followNotification.getInterfaces().map(({ name }) => name),
-    ['Node', 'Notification'],
-  );
-  assert.equal(String(followNotification.getFields().profile.type), 'Profile!');
-  assert.equal(followNotification.getFields().relatedProfile, undefined);
-  assert.equal(followNotification.getFields().kind, undefined);
-  assert.equal(followNotification.getFields().sourceId, undefined);
-  assert.equal(followNotification.getFields().data, undefined);
-  assert.equal(schema.getType('NotificationType'), undefined);
-
-  assert.ok(isObjectType(reactionNotification));
-  assert.deepEqual(
-    reactionNotification.getInterfaces().map(({ name }) => name),
-    ['Node', 'Notification'],
-  );
-  assert.equal(String(reactionNotification.getFields().profile.type), 'Profile!');
-  assert.equal(String(reactionNotification.getFields().post.type), 'Post!');
-  assert.equal(String(reactionNotification.getFields().type.type), 'String!');
-  assert.equal(reactionNotification.getFields().kind, undefined);
-  assert.equal(reactionNotification.getFields().sourceId, undefined);
-  assert.equal(reactionNotification.getFields().data, undefined);
-
-  assert.ok(isObjectType(replyNotification));
-  assert.deepEqual(
-    replyNotification.getInterfaces().map(({ name }) => name),
-    ['Node', 'Notification'],
-  );
-  assert.equal(String(replyNotification.getFields().profile.type), 'Profile!');
-  assert.equal(String(replyNotification.getFields().post.type), 'Post!');
-  assert.equal(replyNotification.getFields().kind, undefined);
-  assert.equal(replyNotification.getFields().sourceId, undefined);
-  assert.equal(replyNotification.getFields().data, undefined);
-
-  assert.ok(isObjectType(repostNotification));
-  assert.deepEqual(
-    repostNotification.getInterfaces().map(({ name }) => name),
-    ['Node', 'Notification'],
-  );
-  assert.equal(String(repostNotification.getFields().profile.type), 'Profile!');
-  assert.equal(String(repostNotification.getFields().post.type), 'Post!');
-  assert.equal(repostNotification.getFields().kind, undefined);
-  assert.equal(repostNotification.getFields().sourceId, undefined);
-  assert.equal(repostNotification.getFields().data, undefined);
-
-  assert.ok(isInterfaceType(notification));
-  assert.deepEqual(
-    notification.getInterfaces().map(({ name }) => name),
-    ['Node'],
-  );
-  assert.equal(notificationNodeType('FOLLOW'), 'FollowNotification');
-  assert.equal(notificationNodeType('REACTION'), 'ReactionNotification');
-  assert.equal(notificationNodeType('REPLY'), 'ReplyNotification');
-  assert.equal(notificationNodeType('REPOST'), 'RepostNotification');
-  assert.equal(notificationNodeType('UNSUPPORTED'), null);
-  assert.equal(String(profile.getFields().notifications?.type), 'NotificationConnection!');
-  assert.equal(String(profile.getFields().unreadNotificationCount?.type), 'Int!');
-
-  const connection = schema.getType('NotificationConnection');
-  const edge = schema.getType('NotificationConnectionEdge');
-  assert.ok(isObjectType(connection));
-  assert.ok(isObjectType(edge));
-  assert.equal(String(connection.getFields().pageInfo.type), 'PageInfo!');
-  assert.equal(String(edge.getFields().cursor.type), 'String!');
-  assert.equal(String(edge.getFields().node.type), 'Notification!');
-});
-
-test('exposes the typed Notification Read mutation payload', () => {
-  const mutation = schema.getMutationType();
-  const input = schema.getType('MarkNotificationReadInput');
-  const payload = schema.getType('MarkNotificationReadPayload');
-
-  assert.equal(
-    String(mutation?.getFields().markNotificationRead?.type),
-    'MarkNotificationReadPayload!',
-  );
-  assert.ok(isInputObjectType(input));
-  assert.equal(String(input.getFields().id.type), 'ID!');
-  assert.ok(isObjectType(payload));
-  assert.equal(String(payload.getFields().notification.type), 'Notification!');
-  assert.equal(String(payload.getFields().recipientProfile.type), 'Profile!');
-});
-
-test('exposes current-session revoke without a target input', () => {
-  const mutation = schema.getMutationType();
-  const payload = schema.getType('RevokeCurrentSessionPayload');
-  const field = mutation?.getFields().revokeCurrentSession;
-
-  assert.equal(String(field?.type), 'RevokeCurrentSessionPayload!');
-  assert.deepEqual(field?.args, []);
-  assert.ok(isObjectType(payload));
-  assert.deepEqual(Object.keys(payload.getFields()), ['completed']);
-  assert.equal(String(payload.getFields().completed.type), 'Boolean!');
-});
-
-test('exposes the private Bookmark Node and Profile connection contract', () => {
-  const bookmark = schema.getType('Bookmark');
-  const profile = schema.getType('Profile');
-  const connection = schema.getType('BookmarkConnection');
-  const edge = schema.getType('BookmarkConnectionEdge');
-
-  assert.ok(isObjectType(bookmark));
-  assert.ok(isObjectType(profile));
-  assert.deepEqual(
-    bookmark.getInterfaces().map(({ name }) => name),
-    ['Node'],
-  );
-  assert.equal(String(bookmark.getFields().createdAt.type), 'DateTime!');
-  assert.equal(String(bookmark.getFields().post.type), 'Post');
-  assert.equal(String(bookmark.getFields().profile.type), 'Profile!');
-  assert.equal(bookmark.getFields().postId, undefined);
-  assert.equal(String(profile.getFields().bookmarks?.type), 'BookmarkConnection!');
-
-  assert.ok(isObjectType(connection));
-  assert.ok(isObjectType(edge));
-  assert.equal(String(connection.getFields().pageInfo.type), 'PageInfo!');
-  assert.equal(String(edge.getFields().cursor.type), 'String!');
-  assert.equal(String(edge.getFields().node.type), 'Bookmark!');
-});
-
-test('rejects legacy raw UUID and unknown typename Node IDs', async () => {
-  for (const id of [
-    '00000000-0000-8006-8000-000000000001',
-    encodeGlobalId('Unknown', '00000000-0000-8006-8000-000000000001'),
-  ]) {
-    const result = await graphql({
-      schema,
-      source: `query Node($id: ID!) { node(id: $id) { id } }`,
-      variableValues: { id },
-      contextValue: {},
-    });
-
-    assert.equal(result.data?.node, null);
-    assert.ok(result.errors?.[0]);
-  }
-});
-
-test('targets the selected Profile without accepting a Profile ID in update input', () => {
-  const input = schema.getType('UpdateProfileInput');
-
-  assert.ok(isInputObjectType(input));
-  assert.deepEqual(Object.keys(input.getFields()).sort(), [
-    'bio',
-    'displayName',
-    'followPolicy',
-    'tags',
-  ]);
-});
-
-test('exposes Profile tags as Hashtag nodes', () => {
-  const hashtag = schema.getType('Hashtag');
-  const profile = schema.getType('Profile');
-
-  assert.ok(isObjectType(hashtag));
-  assert.ok(isObjectType(profile));
-  assert.deepEqual(
-    hashtag.getInterfaces().map(({ name }) => name),
-    ['Node'],
-  );
-  assert.equal(String(hashtag.getFields().id.type), 'ID!');
-  assert.equal(String(hashtag.getFields().name.type), 'String!');
-  assert.equal(String(profile.getFields().tags.type), '[Hashtag!]!');
-});
-
-test('rejects a non-Notification global ID for Notification Read', async () => {
-  const result = await graphql({
-    schema,
-    source: `mutation MarkNotificationRead($id: ID!) {
-      markNotificationRead(input: { id: $id }) { notification { id } }
-    }`,
-    variableValues: {
-      id: encodeGlobalId('Profile', '00000000-0000-8006-8000-000000000001'),
-    },
-    contextValue: { session: { accountId: 'account', id: 'session' } },
-  });
-
-  assert.equal(result.data, null);
-  assert.match(result.errors?.[0]?.message ?? '', /Notification not found/);
-});
